+++ src-tauri/src/book/mod.rs
@@ -1,0 +1,2 @@
+pub mod bookio;
+pub mod util;
