+++ src-tauri/src/lib.rs
@@ -1,5 +1,4 @@
-
 pub mod book;
+pub mod book_item;
 pub mod shelf;
 pub mod xml;
-
